Create README.md
As the first cryptocurrency project, the COE project is developing a community-oriented economic flow on the ICP blockchain. COE operates as a large DAO and hopes to make a good impact in the global community and help people's financial freedom. The COE project consists of three NFT collections and a decentralized token. In the second phase of the project roadmap, the COE will develop traditional businesses to direct the financial flow from the traditional market to its decentralized token structure.